Refactor: Move elements into specific folders, add analysis options and update code to meet alaysis requirements
