+++ README.md
@@ -1,17 +1,8 @@
-# flutter_kcal
+# Kcal
 
-A new Flutter project.
+Kcal is a nutrition app built using Flutter.
 
-## Getting Started
+The designs for this project are by Salmanwap on UI8.
 
-This project is a starting point for a Flutter application.
-
-A few resources to get you started if this is your first Flutter project:
-
-- [Lab: Write your first Flutter app](https://flutter.dev/docs/get-started/codelab)
-- [Cookbook: Useful Flutter samples](https://flutter.dev/docs/cookbook)
-
-For help getting started with Flutter, view our
-[online documentation](https://flutter.dev/docs), which offers tutorials,
-samples, guidance on mobile development, and a full API reference.
-# flutter_kcal
+- [Salmanwap UI8 Profile](https://ui8.net/salmanwap)
+- [Kcal UI8 Profile](https://ui8.net/salmanwap/products/kcal---nutrition-app-ui-kit)
