+++ analysis_options.yaml
@@ -1,0 +1,11 @@
+include: package:very_good_analysis/analysis_options.yaml
+
+analyzer:
+  exclude:
+    - lib/**.*.dart
+    - test/**.*.dart
+
+linter:
+  rules:
+    public_member_api_docs: true
+    prefer_relative_imports: false
