+++ lib/splash/presentation/splash_page.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 
+/// SplashPage
 class SplashPage extends StatelessWidget {
+  /// Initializes [key] for subclasses.
   const SplashPage({Key? key}) : super(key: key);
 
   @override
