+++ lib/theme/core/colors.dart
@@ -1,0 +1,49 @@
+import 'dart:ui';
+
+/// Primary Color
+const kPrimaryColor = Color(0xFF91C788);
+
+/// Secondary Color
+const kSecondaryColor = Color(0xFFFF9385);
+
+/// Primary Accent Color
+const kPrimaryAccent = Color(0xFFFFD885);
+
+/// Secondary Accent Color
+const kSecondaryAccent = Color(0xFF76BFFF);
+
+/// Tertiary Accent Color
+const kTertiaryAccent = Color(0xFF9E9BC7);
+
+/// Typography Title Color
+const kTypographyTitle = Color(0xFF0E140E);
+
+/// Typography Subtitle Color
+const kTypographySubtitle = Color(0xFF3E433E);
+
+/// Typography Body Color
+const kTypographyBody = Color(0xFF565B56);
+
+/// Typography Caption Color
+const kTypographyCaption = Color(0xFF6E726E);
+
+/// Typography White Color
+const kTypographyWhite = Color(0xFFFFFFFF);
+
+/// Borders 900 Color
+const kBorders900 = Color(0xFF6E726E);
+
+/// Borders 700 Color
+const kBorders700 = Color(0xFF878A87);
+
+/// Borders 500 Color
+const kBorders500 = Color(0xFF9FA19F);
+
+/// Borders 300 Color
+const kBorders300 = Color(0xFFCFD0CF);
+
+/// Borders 100 Color
+const kBorders100 = Color(0xFFE7E8E7);
+
+/// Borders 50 Color
+const kBorders50 = Color(0xFFF7F8F7);
