+++ test/splash/presentation/splash_page_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_kcal/splash/presentation/splash_page.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:kcal/splash/presentation/splash_page.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    await tester.pumpWidget(SplashPage());
+    await tester.pumpWidget(const SplashPage());
 
     final splashImageFinder = find.byType(Image);
